POSIX::mblen() Allow not resetting shift state
This enhances this function so that when the first parameter is undef,
it resets any shift state, and when the parameter is non-null the state
is retained.  Previously the state was always reset with every call,
preventing adequate use of this in locales (very rarely encountered)
that have shift states.

This commit also changes it so that mblen is used instead of mbrlen if
operating under thread-safe locales.  This just made the coding easier.
And, unless there are bugs in the C library implementation of mblen,
should not matter.

[perl #133928] showed that someone was having trouble with shift states.