POSIX::mbtowc, wctomb: Allow not resetting shift state
This commit enhances these functions so that when the appropriate
parameter is undef, any shift state is reset, and when the parameter is
non-null the state is retained.  Previously the state was always reset
with every call, preventing adequate use of these in locales (very
rarely encountered) that have shift states.

This commit also changes it so that mbtowc and wctomb are used instead
of mbrtowc and wcrtomb if operating under thread-safe locales.  This
just made the coding easier.  And, unless there are bugs in the C
library implementations, should not matter.